requirements-airflow.txt
pyspark==3.5.1
pandas==2.2.2
pyarrow==15.0.2
sqlalchemy==2.0.32
psycopg2-binary==2.9.9
scikit-learn==1.8.0
mlflow==2.14.1
